backend/requirements.txt
# VisionX ML Backend - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# Model Serialization
joblib==1.3.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# MLOps
mlflow==2.10.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Explainability
shap==0.44.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Performance
orjson==3.9.12